debug: simplify to basic Vercel Python function
- Use basic Python HTTP handler to test Vercel function detection
- Move FastAPI code to main.py temporarily
- Simplify vercel.json configuration
- This will help diagnose if Vercel can detect Python functions at all
